section 25 done

+++ Node.js/restAPI/controllers/feedController.js
@@ -1,11 +1,14 @@
 const feedController = require("express").Router();
 const createPostSchema = require("../validations/postSchema");
 const postModel = require("../models/post");
+const userModel = require("../models/user");
 const clearImage = require("../utils/clearImage");
+const isAuth = require("../middleware/isAuth");
+const ValidationError = require("../utils/createValidationError");
 
 const ITEMS_PER_PAGE = 2;
 
-feedController.get("/getPosts", async (req, res, next) => {
+feedController.get("/getPosts", isAuth, async (req, res, next) => {
   try {
     let page = Number(req.query.page);
     page < 1 || isNaN(page) ? (page = 1) : page;
@@ -13,7 +16,7 @@ feedController.get("/getPosts", async (req, res, next) => {
     const totalPosts = await postModel.find().countDocuments();
     const totalPages = Math.ceil(totalPosts / ITEMS_PER_PAGE);
 
-    page > totalPages ? (page = totalPages) : page;
+    page = totalPages !== 0 ? Math.min(page, totalPages) : 1;
 
     const posts = await postModel
       .find()
@@ -26,17 +29,21 @@ feedController.get("/getPosts", async (req, res, next) => {
   }
 });
 
-feedController.post("/createPost", async (req, res, next) => {
+feedController.post("/createPost", isAuth, async (req, res, next) => {
   try {
     const { title, content } = createPostSchema.partial().parse(req.body);
     if (!req.file) {
       throw new Error({ message: "No image provided", statusCode: 422 });
     }
     const imageUrl = req.file.path;
-    const newPost = await postModel.create({ title, content, imageUrl, creator: { name: "Nik" } });
+
+    const newPost = await postModel.create({ title, content, imageUrl, creator: req.userId });
+    const user = await userModel.findByIdAndUpdate({ _id: req.userId }, { $push: { posts: newPost } }, { new: true });
+
     res.status(201).json({
       message: "Post created successfully",
       post: newPost,
+      creator: { _id: user._id, name: user.name },
     });
   } catch (err) {
     err.statusCode = 422;
@@ -44,7 +51,7 @@ feedController.post("/createPost", async (req, res, next) => {
   }
 });
 
-feedController.get("/post/:postId", async (req, res, next) => {
+feedController.get("/post/:postId", isAuth, async (req, res, next) => {
   try {
     const { postId } = req.params;
     const post = await postModel.findById(postId);
@@ -57,7 +64,7 @@ feedController.get("/post/:postId", async (req, res, next) => {
   }
 });
 
-feedController.put("/post/:postId", async (req, res, next) => {
+feedController.put("/post/:postId", isAuth, async (req, res, next) => {
   try {
     const { postId } = req.params;
     const { title, content } = createPostSchema.partial().parse(req.body);
@@ -69,6 +76,11 @@ feedController.put("/post/:postId", async (req, res, next) => {
       throw new Error({ message: "No image provided", statusCode: 422 });
     }
     const post = await postModel.findById(postId);
+
+    if (post.creator.toString() !== req.userId) {
+      throw new ValidationError({ message: "Not authorized", statusCode: 403 });
+    }
+
     if (imageUrl !== post.imageUrl) {
       clearImage(post.imageUrl);
     }
@@ -79,17 +91,24 @@ feedController.put("/post/:postId", async (req, res, next) => {
   }
 });
 
-feedController.delete("/post/:postId", async (req, res, next) => {
+feedController.delete("/post/:postId", isAuth, async (req, res, next) => {
   try {
     const { postId } = req.params;
     const post = await postModel.findById(postId);
+
     if (!post) {
-      throw new Error({ message: "Could not find specific post", statusCode: 404 });
+      throw new ValidationError({ message: "Could not find specific post", statusCode: 404 });
+    }
+
+    if (post.creator.toString() !== req.userId) {
+      throw new ValidationError({ message: "Not authorized", statusCode: 403 });
     }
 
     clearImage(post.imageUrl);
 
     await postModel.findByIdAndDelete(postId);
+    await userModel.findByIdAndUpdate(req.userId, { $pull: { posts: post._id } });
+
     res.status(200).json({ message: "Post deleted." });
   } catch (err) {
     next(err);
